Also had one more question from codcad (Recursion fever) just by changing from int to long long (I'm so dumb I took 3 days to realize it)

+++ codcad/dp/recursion_fever.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 #define MAXN 1100
 
-int a[MAXN], f[MAXN];
+long long a[MAXN], f[MAXN];
 
 int main(){
     int n, k, m;
